商品服务-录入商品 1.保存spu基本信息 pms_spu_info 2.保存spu的描述图片 pms_spu_info_desc 3.保存spu的图片集 pms_spu_images 4.保存spu的规格参数 pms_product_attr_value 5.保存spu的积分信息 gulimall_sms -> sms_spu_bounds 6.保存当前spu对应的所有sku信息 6.1 sku的基本信息 pms_sku_info 6.2 sku的图片信息 pms_spu_images 6.3 sku的销售属性信息 pms_sku_sale_attr_value 6.4 sku的优惠满减信息 gulimall_sms

+++ gulimall-product/src/main/java/com/tech/gulimall/product/entity/vo/Skus.java
@@ -44,11 +44,11 @@ public class Skus {
     /**
      * 满几件
      */
-    private BigDecimal fullCount;
+    private int fullCount;
     /**
      * 打几折
      */
-    private int discount;
+    private BigDecimal discount;
     /**
      * 是否叠加其他优惠[0-不可叠加，1-可叠加]
      */
